GRD-38770: Filter to set time/stamp

+++ logstash-filter-mongodb-guardium/src/main/java/com/ibm/guardium/connector/structures/Data.java
@@ -4,7 +4,7 @@ import com.ibm.guardium.Construct;
 
 public class Data {
     private Construct construct;
-    private int timestamp;
+    private int timestamp; // limit warning: use new Date((long) unixTime * 1000), to revert to Date, otherwise, int max will be reached.
     private String originalSqlCommand;
     private boolean useConstruct;
 
@@ -28,7 +28,7 @@ public class Data {
         this.construct = construct;
     }
 
-    public int getTimestamp() {
+    public long getTimestamp() {
         return timestamp;
     }
